# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tole Community Management API",
    "version": "1.0.0",
...
    "contact": {
      "name": "Tole Community Management System",
      "url": "https://github.com/"
    },
    "license": {
      "name": "UNLICENSED"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "This server — all versioned endpoints live under the /api prefix."
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste the `token` returned by `POST /auth/login` into the Authorize dialog. It is sent as `Authorization: Bearer <token>`. Query-string tokens are rejected by the server."
      }
    },
    "schemas": {
...
      "NoticeType": {
        "type": "string",
        "enum": [
          "general",
          "emergency",
          "event",
          "maintenance"
        ],
        "description": "Notice classification; also selects the emoji prefix on the pushed notification."
      },
...
      "Notice": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "65f1a2b3c4d5e6f7a8b9c0d6"
          },
          "title": {
            "type": "string",
            "minLength": 5,
            "maxLength": 100,
            "example": "Water supply interruption on Friday"
          },
          "content": {
            "type": "string",
            "minLength": 10,
            "maxLength": 1000,
            "example": "Supply will be interrupted from 9 AM to 2 PM for tank maintenance."
          },
          "type": {
            "type": "string",
            "enum": [
              "general",
              "emergency",
              "event",
              "maintenance"
            ],
            "description": "Notice classification; also selects the emoji prefix on the pushed notification."
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2026-12-31T00:00:00.000Z",
            "description": "Defaults to 7 days after creation when omitted."
          },
          "expired": {
            "type": "boolean",
            "example": false,
            "description": "True when the notice expiry has passed. Included on notice list responses."
          },
          "targetSections": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "Section 1"
            ],
            "description": "Empty array = broadcast to every section. Residents only ever see notices matching their own section(s) or broadcasts."
          },
          "createdBy": {
            "type": "object",
            "description": "Populated with `name` only.",
            "properties": {
              "_id": {
                "type": "string",
                "example": "65f1a2b3c4d5e6f7a8b9c0d1"
              },
              "name": {
                "type": "string",
                "minLength": 2,
                "maxLength": 50,
                "pattern": "^[A-Za-z][A-Za-z ]*$",
                "example": "Ram Bahadur"
              },
              "username": {
                "type": "string",
                "example": "ram.bahadur"
              },
              "phone": {
                "type": "string",
                "pattern": "^(?:98[012456]|97[01456])\\d{7}$",
                "example": "9841234567",
                "description": "Optional. A 10-digit Nepal mobile number starting with 980, 981, 982, 984, 985, 986, 970, 971, 974, 975, or 976."
              },
              "email": {
                "type": "string",
                "format": "email",
                "maxLength": 100,
                "example": "ram.bahadur@example.com"
              },
              "specialization": {
                "type": "string",
                "nullable": true,
                "enum": [
                  "water",
                  "electric",
                  "lift",
                  "sanitation",
                  "security",
                  "general",
                  null
                ],
                "description": "Only populated in projections that select it."
              }
            }
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Deletion sets this to `false` (soft delete) rather than removing the row."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateNoticeRequest": {
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 5,
            "maxLength": 100,
            "example": "Water supply interruption on Friday"
          },
          "content": {
            "type": "string",
            "minLength": 10,
            "maxLength": 1000,
            "example": "Supply will be interrupted from 9 AM to 2 PM for tank maintenance."
          },
          "type": {
            "type": "string",
            "enum": [
              "general",
              "emergency",
              "event",
              "maintenance"
            ],
            "description": "Notice classification; also selects the emoji prefix on the pushed notification.",
            "default": "general"
          },
          "targetSections": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "Section 1"
            ],
            "description": "Optional. Omit or pass `[]` to broadcast to everyone. Every entry must match an existing house section, otherwise 400."
          },
          "expiresAt": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2026-12-31",
            "description": "Optional expiry date; if supplied it must be valid and in the future. Omit or leave empty for no expiry."
          }
        }
      },
...
      "ValidationError": {
        "type": "object",
        "description": "A single failed field check, as produced by `express-validator`.",
        "properties": {
          "field": {
            "type": "string",
            "example": "email"
          },
          "message": {
            "type": "string",
            "example": "Invalid email format"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Access denied"
          },
          "code": {
            "type": "string",
            "example": "RATE_LIMITED",
            "description": "Machine-readable error code when provided."
          }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "message",
          "errors"
        ],
        "description": "The `express-validator` failure envelope. Note the contrast with the export-filter error, whose `errors` array contains plain strings instead of `{ field, message }` objects.",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
    },
    "parameters": {
      "PageParam": {
        "name": "page",
        "in": "query",
        "required": false,
        "description": "Page number, 1-based. **Omit this entirely to disable pagination** and receive the full result set — several frontend screens (house pickers, section pickers) rely on that behaviour.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1,
          "example": 1
        }
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Items per page. Clamped to the range 1-100. Has no effect unless `page` is also supplied.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20,
          "example": 20
        }
      },
...
      "IdPathParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "MongoDB ObjectId of the resource.",
        "schema": {
          "type": "string",
          "pattern": "^[0-9a-fA-F]{24}$",
          "example": "65f1a2b3c4d5e6f7a8b9c0d1"
        }
      },
...
      "SearchQuery": {
        "name": "search",
        "in": "query",
        "required": false,
        "description": "Case-insensitive substring match. The value is regex-escaped, so it is matched literally. Must be 100 characters or fewer — a longer value returns 400.",
        "schema": {
          "type": "string",
          "maxLength": 100,
          "example": "water"
        }
      },
...
      "NoticeTypeQuery": {
        "name": "type",
        "in": "query",
        "required": false,
        "description": "Filter notices by type.",
        "schema": {
          "$ref": "#/components/schemas/NoticeType"
        }
      },
      "IncludeExpiredQuery": {
        "name": "includeExpired",
        "in": "query",
        "required": false,
        "description": "Admins and staff may pass `true` to include expired active notices. Residents always receive unexpired notices only.",
        "schema": {
          "type": "string",
          "enum": [
            "true",
            "false"
          ],
          "example": "true"
        }
      },
...
    },
    "responses": {
      "UnauthorizedResponse": {
        "description": "Authentication failed or was not supplied. Returned when the `Authorization` header is missing, the token is invalid or expired, the referenced user no longer exists, the account is deactivated, or a token was passed as a `token` / `access_token` **query parameter** (which this API rejects on purpose to avoid leaking credentials in logs).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "examples": {
          "missingToken": {
            "summary": "No Authorization header",
            "value": {
              "success": false,
              "message": "Not authorized"
            }
          },
          "queryToken": {
            "summary": "Token supplied in the query string",
            "value": {
              "success": false,
              "message": "Use the Authorization request header"
            }
          },
          "inactive": {
            "summary": "Deactivated account",
            "value": {
              "success": false,
              "message": "User account is inactive or not found"
            }
          },
          "invalid": {
            "summary": "Malformed or bad signature",
            "value": {
              "success": false,
              "message": "Token invalid"
            }
          }
        }
      },
      "ForbiddenResponse": {
        "description": "The caller is authenticated but not permitted. Covers role mismatches (`authorize`), record-level ownership checks (a resident touching another resident's due or complaint), and denied export attempts.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "examples": {
          "role": {
            "summary": "Wrong role for this endpoint",
            "value": {
              "success": false,
              "message": "Access denied"
            }
          },
          "export": {
            "summary": "Export not permitted for this staff member",
            "value": {
              "success": false,
              "message": "You are not allowed to export dues data."
            }
          }
        }
      },
      "NotFoundResponse": {
        "description": "The resource does not exist. Also returned when a file is requested that is missing on disk, and for malformed ObjectIds (`errorHandler` maps Mongoose `CastError` to 404).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "BadRequestResponse": {
        "description": "The request was rejected. Covers business-rule violations, invalid enum values, rejected uploads (`INVALID_FILE_TYPE`, `LIMIT_FILE_SIZE`), Mongoose validation errors and duplicate-key errors.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ValidationFailedResponse": {
        "description": "One or more fields failed `express-validator` checks. Any files uploaded alongside the request are deleted before responding.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationErrorResponse"
            }
          }
        },
        "example": {
          "success": false,
          "message": "Validation failed",
          "errors": [
            {
              "field": "email",
              "message": "Invalid email format"
            },
            {
              "field": "password",
              "message": "Password must be at least 6 characters"
            }
          ]
        }
      },
...
      "ServerErrorResponse": {
        "description": "Unhandled server error. `message` is the raw error text, so treat it as diagnostic only.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/notices": {
      "get": {
        "operationId": "listNotices",
        "summary": "List notices",
        "description": "Returns active, unexpired notices, newest first. `search` matches the title or content case-insensitively (regex-escaped, max 100 characters).\n\n**Residents only see notices addressed to their own section**, plus any notice with an empty `targetSections` array (a broadcast). Residents never receive expired notices. Admins and staff may pass `includeExpired=true` to include expired active notices; those rows have `expired: true`. Expired notices are retained and never auto-deleted.\n\n**Pagination is opt-in** — omit `page` to receive the full list, which the frontend section pickers depend on.",
        "tags": [
          "Notices"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SearchQuery"
          },
          {
            "$ref": "#/components/parameters/NoticeTypeQuery"
          },
          {
            "$ref": "#/components/parameters/IncludeExpiredQuery"
          },
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Matching notices with pagination metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer",
                      "example": 10
                    },
                    "total": {
                      "type": "integer",
                      "example": 34
                    },
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "pages": {
                      "type": "integer",
                      "example": 4
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Notice"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "500": {
            "$ref": "#/components/responses/ServerErrorResponse"
          }
        }
      },
      "post": {
        "operationId": "createNotice",
        "summary": "Create a notice and notify residents",
        "description": "Publishes a notice and immediately pushes an in-app notification to the affected residents.\n\nAn optional `expiresAt` must be a valid date in the future; when omitted or empty it defaults to 7 days after creation.\n\n**Targeting:** omit `targetSections` (or send `[]`) to notify every active resident. When sections are supplied, every entry must match an existing house section — an unknown name returns 400 — and only residents linked to a house in those sections are notified.\n\nThe notification step is best-effort: the notice is saved even if delivery fails, in which case `notificationDelivered` is `false` and a `warning` explains why.",
        "tags": [
          "Notices"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoticeRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "The notice was created and notifications were dispatched.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Notice"
                    },
                    "notifiedCount": {
                      "type": "integer",
                      "example": 37,
                      "description": "Number of residents notified."
                    },
                    "notificationDelivered": {
                      "type": "boolean",
                      "example": true
                    },
                    "warning": {
                      "type": "string",
                      "nullable": true,
                      "example": "Notice was saved, but notifications could not be delivered."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailedResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "500": {
            "$ref": "#/components/responses/ServerErrorResponse"
          }
        }
      }
    },
    "/notices/{id}/expire": {
      "put": {
        "operationId": "expireNotice",
        "summary": "Expire a notice immediately",
        "description": "Sets `expiresAt` to the current time. Admins and staff may expire notices; the notice then disappears from the default active-notices list.",
        "tags": [
          "Notices"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPathParam"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The notice was expired.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Notice"
                    },
                    "message": {
                      "type": "string",
                      "example": "Notice expired successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/ServerErrorResponse"
          }
        }
      }
    },
...
  }
}